Adapt tests of huggingface flax model
