+++ CI/integration_tests/models/test_huggingface_flax_model.py
@@ -25,21 +25,24 @@ Summary
 -------
 """
 
+import os
+
+os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
+
 import jax.numpy as np
+import numpy as np
 import optax
-import pytest
-from flax import linen as nn
 from jax import random
 from transformers import FlaxResNetForImageClassification, ResNetConfig
 
+from znnl.loss_functions import MeanPowerLoss
 from znnl.models import HuggingFaceFlaxModel
+from znnl.training_strategies import SimpleTraining
 
 
 class TestFlaxHFModule:
     """
-    Integration test suite for the flax Hugging Face (HF) module.
-
-    Train the model for a few steps and check whether the loss decreases.
+    Test suite for the flax Hugging Face (HF) module.
     """
 
     @classmethod
@@ -49,7 +52,7 @@ class TestFlaxHFModule:
         The resnet config has a 1 dimensional input and a 2 dimensional output.
         """
 
-        resnet_config = ResNetConfig(
+        resnet50_config = ResNetConfig(
             num_channels=2,
             embedding_size=64,
             hidden_sizes=[256, 512, 1024, 2048],
@@ -59,34 +62,80 @@ class TestFlaxHFModule:
             downsample_in_first_stage=False,
             out_features=None,
             out_indices=None,
-            id2label=dict(zip([1, 2], [1, 2])),
+            id2label={
+                i: i for i in range(3)
+            },  # Dummy labels to define the output dimension
             return_dict=True,
         )
-        hf_model = FlaxResNetForImageClassification(
-            config=resnet_config,
+
+        # ResNet-18 taken from https://huggingface.co/microsoft/resnet-18/blob/main/config.json
+        resnet18_config = ResNetConfig(
+            num_channels=2,
+            embedding_size=64,
+            hidden_sizes=[64, 128, 256, 512],
+            depths=[2, 2, 2, 2],
+            layer_type="basic",
+            hidden_act="relu",
+            downsample_in_first_stage=False,
+            id2label={
+                i: i for i in range(3)
+            },  # Dummy labels to define the output dimension
+            return_dict=True,
+        )
+
+        resnet18 = FlaxResNetForImageClassification(
+            config=resnet50_config,
             input_shape=(1, 8, 8, 2),
             seed=0,
             _do_init=True,
         )
-        cls.model = HuggingFaceFlaxModel(
-            hf_model,
-            optax.adam(learning_rate=0.001),
+        resnet50 = FlaxResNetForImageClassification(
+            config=resnet18_config,
+            input_shape=(1, 8, 8, 2),
+            seed=0,
+            _do_init=True,
+        )
+        cls.resnet18 = HuggingFaceFlaxModel(
+            resnet18,
+            optax.sgd(learning_rate=1e-3),
+            batch_size=3,
+        )
+        cls.resnet50 = HuggingFaceFlaxModel(
+            resnet50,
+            optax.sgd(learning_rate=1e-3),
             batch_size=3,
         )
 
         key = random.PRNGKey(0)
-        cls.x = random.normal(key, (3, 2, 8, 8))
+        cls.train_ds = {
+            "inputs": random.normal(key, (30, 2, 8, 8)),
+            "targets": np.repeat(np.eye(3), 10, axis=0),
+        }
 
-    def test_ntk_shape(self):
+    def train_model(self, model):
         """
-        Test whether the NTK shape is correct.
+        Train the model for a few steps.
         """
-        ntk = self.model.compute_ntk(self.x)["empirical"]
-        assert ntk.shape == (3, 3)
+        trainer = SimpleTraining(
+            model=model,
+            loss_fn=MeanPowerLoss(order=2),
+        )
+        batched_loss = trainer.train_model(
+            train_ds=self.train_ds,
+            test_ds=self.train_ds,
+            epochs=20,
+            batch_size=30,
+        )
+        return batched_loss
 
-    def test_infinite_failure(self):
+    def test_loss_decreaser(self):
         """
-        Test that the call to the infinite NTK fails.
+        Test whether the training loss decreases.
         """
-        with pytest.raises(NotImplementedError):
-            self.model.compute_ntk(self.x, infinite=True)
+        batched_loss = np.array(self.train_model(self.resnet18)["train_losses"])
+        diff = batched_loss[10:] - batched_loss[:1]
+        np.testing.assert_array_less(diff, 0)
+
+        batched_loss = np.array(self.train_model(self.resnet50)["train_losses"])
+        diff = batched_loss[10:] - batched_loss[:1]
+        np.testing.assert_array_less(diff, 0)
